session: in-memory session service implementation (#16)
* session: in-memory session service implementation

And made adk.Session a snapshot of the internal session state
by removing AppendEvent and access to SessionService.

TODO: it's still unclear to me what's the desired semantic between
the (local) Session and the (remote) session state, and how/when
updates on one side should be reflected to the other side.

* session: use ordered key and omap, implement List

This allows simpler internal data structure
by avoiding the deeply nested map.
Rewrote tests.